🔧 refactor: 적금 Data, Domain 계층 model 디폴트값 추가

+++ finafanAndroid/app/src/main/java/com/a702/finafan/domain/savings/model/Star.kt
@@ -1,7 +1,7 @@
 package com.a702.finafan.domain.savings.model
 
 data class Star(
-    val entertainerId: Int = 0,
+    val entertainerId: Long = 0,
     val entertainerName: String = "",
     val birthDate: String = "",
     val entertainerProfileUrl: String = "",
